// frontend/pi-front-end-applications/pi-components-catalog/molecules/src/common/CompanySelection/tableFields.enum.ts
// Fields that are shown in company list modal
// For the table to render correctly the table column keys should match the property names in the row object
// Update the enum to ensure integrity of table columns and rows

export enum TABLE_FIELDS {
  COMPANY_NAME = 'name',
  ADDRESS = 'address',
  TELEPHONE_NUMBER = 'telephoneNumber',
  CORP_ID = 'corpId',
  COMPANY_ID = 'companyId',
  AR_NUMBER = 'arNumber',
  RESTRICTED = 'restricted',
}